requirements.txt
asgiref==3.7.2
Django==4.2.1
djangorestframework==3.14.0
docx==0.2.4
lxml==4.9.2
Pillow==9.5.0
psycopg2==2.9.6
python-docx==0.8.11
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.2
